feat: Implement Responses API passthrough handling
- Added `handleResponsesPassthrough` method to the Server struct to manage minimal patching of requests to the Responses API, preserving original request body fields.
- Introduced `bodyHasInputField` function to quickly check for the presence of the "input" key in JSON bodies.
- Updated the `universal.go` file to route requests with an "input" field to the new passthrough handler.
- Enhanced the `composeInstructionsForRoute` function to handle instruction composition for the Responses API.
- Modified the `Client` struct to include a new `DoRaw` method for sending pre-built JSON payloads to the backend.
- Updated usage structures in `types` to support detailed token usage breakdowns, including input and output token details.
- Added tests to ensure correct behavior of the new passthrough functionality and instruction composition.
- Refactored numeric handling in `helpers.go` to use int64 for token counts, ensuring compatibility with larger values.

+++ internal/types/anthropic.go
@@ -79,8 +79,8 @@ type AnthropicContentOut struct {
 
 // AnthropicUsage holds Messages API usage.
 type AnthropicUsage struct {
-	InputTokens  int `json:"input_tokens"`
-	OutputTokens int `json:"output_tokens"`
+	InputTokens  int64 `json:"input_tokens"`
+	OutputTokens int64 `json:"output_tokens"`
 }
 
 // AnthropicModelListResponse is the response for GET /v1/models in Anthropic mode.
